+++ pabilo_payment_gateway/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': 'Pabilo Payment Gateway for Odoo',
-    'version': '16.0.2.5.0',
+    'version': '16.0.2.6.0',
     'summary': 'Verifica pagos móviles, transferencias y Binance en el POS con Pabilo',
     'description': """
 Pabilo Payment Gateway